+++ js/modules/constants.js
@@ -1,0 +1,36 @@
+const NAMES = [
+  'Jaga-Jaga',
+  'Katyyyyyyya',
+  'Just Anton',
+  'Vika-hot-chika',
+  'Dasha',
+  'Prodam kirpichi',
+  'Ilya',
+  'Anna',
+  'Devo4ki & Kiri11',
+  'Nastyshka',
+  'Nikita',
+  'Sabinin hater'
+];
+
+const DESCRIPTIONS = [
+  'Записываемся на ноготочки Питер/Москва 1000 р./сеанс',
+  'мы с саньком (не проггеры)',
+  'ахаха Сабина токсик',
+  'нюдсы в тг;)',
+  'Челябинск - столица мира',
+  'РТФ - ЧЕМПИОН!!11!!!!!',
+  'Уважаемые патпищеки, м Я у',
+  'Ниважна, что говорят киски за спиной у крыски'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {NAMES, DESCRIPTIONS, MESSAGES};
