+++ js/modules/utils.js
@@ -1,0 +1,9 @@
+const getInt = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  return Math.floor(Math.random() * (upper - lower + 1) + lower);
+};
+
+const getElement = (array) => array[getInt(0, array.length-1)];
+
+export {getInt, getElement};
